fix: resolve state bugs and replace comment placeholders with real logic
LOGIC:
  - Comments after report evaluation now function correctly.

BUGFIXES:
  - State data is now updated and properly deleted as expected.

evaluation/handlers.py:
  - Updated function names for comment handling.
  - Modified `frontend_cb_re_eval_marks_accepted` to act as a submenu for accepting marks, returning the user to the main menu after the action.

evaluation/keyboards.py:
  - Added keyboards for marks acceptance and comment review.

router.py:
  - Updated with new function names and added route for `cb_re_eval_marks_accepted`.

utils.py:
  - Fixed state data update in `re_delete_auth_messages`.

fsm_states.py:
  - Added new states in `REEvaluationStates` for improved logic flow.

main_menu/handlers.py:
  - Fixed state data update logic.

locale:
  - Added new sections to evaluation in both Russian and English.

+++ source/bot/components/reports_evaluation/utils.py
@@ -50,7 +50,7 @@ def re_require_auth(handler):
         if "jury_code" not in state_data:
             # TODO: Lang instead of just text
             await callback_query.answer(
-                text="Authentication  failed.",
+                text="Authentication failed.",
                 reply_markup=types.ReplyKeyboardRemove(),
                 show_alert=True,
             )
@@ -80,10 +80,11 @@ async def re_delete_auth_messages(state: FSMContext, chat_id: int, bot: Bot):
     state_data = await state.get_data()
     msg_ids = state_data.get("re_auth_message_ids", [])
 
-    for msg_id in msg_ids:
-        try:
-            await bot.delete_message(chat_id, msg_id)
-        except TelegramBadRequest:
-            pass
+    if msg_ids:
+        for msg_id in msg_ids:
+            try:
+                await bot.delete_message(chat_id, msg_id)
+            except TelegramBadRequest:
+                pass
 
-    state_data.pop("re_auth_message_ids", None)
+        await state.update_data(re_auth_message_ids=None)
